backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.3.0
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
stripe==9.9.0
